+++ server/backend/database/mongo/indexes.go
@@ -40,6 +40,8 @@ const (
 	ColSnapshots = "snapshots"
 	// ColSyncedSeqs represents the syncedseqs collection in the database.
 	ColSyncedSeqs = "syncedseqs"
+	// ColVersionVectors represents the versionvector collection in the database.
+	ColVersionVectors = "versionvectors"
 )
 
 // Collections represents the list of all collections in the database.
@@ -147,11 +149,28 @@ var collectionInfos = []collectionInfo{
 			Options: options.Index().SetUnique(true),
 		}, {
 			Keys: bsonx.Doc{
+				{Key: "doc_id", Value: bsonx.Int32(1)}, // shard key
+				{Key: "project_id", Value: bsonx.Int32(1)},
+				{Key: "server_seq", Value: bsonx.Int32(1)},
+			},
+		}, {
+			Keys: bsonx.Doc{
 				{Key: "doc_id", Value: bsonx.Int32(1)},
 				{Key: "project_id", Value: bsonx.Int32(1)},
 				{Key: "lamport", Value: bsonx.Int32(1)},
 				{Key: "actor_id", Value: bsonx.Int32(1)},
 			},
+		}},
+	},
+	{
+		name: ColVersionVectors,
+		indexes: []mongo.IndexModel{{
+			Keys: bsonx.Doc{
+				{Key: "project_id", Value: bsonx.Int32(1)},
+				{Key: "doc_id", Value: bsonx.Int32(1)},
+				{Key: "client_id", Value: bsonx.Int32(1)},
+			},
+			Options: options.Index().SetUnique(true),
 		}},
 	},
 }