+++ server/backend/database/mongo/registry.go
@@ -24,15 +24,17 @@ import (
 	"go.mongodb.org/mongo-driver/bson/bsoncodec"
 	"go.mongodb.org/mongo-driver/bson/bsonoptions"
 	"go.mongodb.org/mongo-driver/bson/bsonrw"
+	"google.golang.org/protobuf/proto"
 
+	"github.com/yorkie-team/yorkie/api/converter"
 	"github.com/yorkie-team/yorkie/api/types"
+	api "github.com/yorkie-team/yorkie/api/yorkie/v1"
 	"github.com/yorkie-team/yorkie/pkg/document/time"
-	"github.com/yorkie-team/yorkie/server/backend/database"
 )
 
 var tID = reflect.TypeOf(types.ID(""))
 var tActorID = reflect.TypeOf(&time.ActorID{})
-var tClientDocInfoMap = reflect.TypeOf(make(database.ClientDocInfoMap))
+var tVersionVector = reflect.TypeOf(time.VersionVector{})
 
 // NewRegistryBuilder returns a new registry builder with the default encoder and decoder.
 func NewRegistryBuilder() *bsoncodec.RegistryBuilder {
@@ -42,16 +44,17 @@ func NewRegistryBuilder() *bsoncodec.RegistryBuilder {
 	bsoncodec.DefaultValueDecoders{}.RegisterDefaultDecoders(rb)
 	bson.PrimitiveCodecs{}.RegisterPrimitiveCodecs(rb)
 
-	// Register the decoder for ObjectID
+	// Register the decoders for types.ID.
 	rb.RegisterCodec(
 		tID,
 		bsoncodec.NewStringCodec(bsonoptions.StringCodec().SetDecodeObjectIDAsHex(true)),
 	)
+	rb.RegisterTypeDecoder(tVersionVector, bsoncodec.ValueDecoderFunc(versionVectorDecoder))
 
-	// Register the encoder for types.ID
+	// Register the encoders for types.ID and time.ActorID.
 	rb.RegisterTypeEncoder(tID, bsoncodec.ValueEncoderFunc(idEncoder))
-	// Register the encoder for time.ActorID
 	rb.RegisterTypeEncoder(tActorID, bsoncodec.ValueEncoderFunc(actorIDEncoder))
+	rb.RegisterTypeEncoder(tVersionVector, bsoncodec.ValueEncoderFunc(versionVectorEncoder))
 
 	return rb
 }
@@ -78,5 +81,55 @@ func actorIDEncoder(_ bsoncodec.EncodeContext, vw bsonrw.ValueWriter, val reflec
 	if err := vw.WriteObjectID(objectID); err != nil {
 		return fmt.Errorf("encode error: %w", err)
 	}
+	return nil
+}
+
+func versionVectorEncoder(_ bsoncodec.EncodeContext, vw bsonrw.ValueWriter, val reflect.Value) error {
+	if !val.IsValid() || val.Type() != tVersionVector {
+		return bsoncodec.ValueEncoderError{Name: "versionVectorEncoder", Types: []reflect.Type{tVersionVector}, Received: val}
+	}
+
+	pbChangeVector, err := converter.ToVersionVector(val.Interface().(time.VersionVector))
+	if err != nil {
+		return err
+	}
+
+	bytes, err := proto.Marshal(pbChangeVector)
+	if err != nil {
+		return fmt.Errorf("encode error: %w", err)
+	}
+
+	if err := vw.WriteBinary(bytes); err != nil {
+		return fmt.Errorf("encode error: %w", err)
+	}
+
+	return nil
+}
+
+func versionVectorDecoder(_ bsoncodec.DecodeContext, vr bsonrw.ValueReader, val reflect.Value) error {
+	if val.Type() != tVersionVector {
+		return bsoncodec.ValueDecoderError{Name: "versionVectorDecoder", Types: []reflect.Type{tVersionVector}, Received: val}
+	}
+
+	switch vrType := vr.Type(); vrType {
+	case bson.TypeBinary:
+		data, _, err := vr.ReadBinary()
+		if err != nil {
+			return fmt.Errorf("decode error: %w", err)
+		}
+
+		var pbVector api.VersionVector
+		if err := proto.Unmarshal(data, &pbVector); err != nil {
+			return fmt.Errorf("decode error: %w", err)
+		}
+
+		vector, err := converter.FromVersionVector(&pbVector)
+		if err != nil {
+			return err
+		}
+
+		val.Set(reflect.ValueOf(vector))
+	}
+
 	return nil
 }